Fixed case where expanding result set nodes through the GUI would bypass custom authorization checks
