+++ engine/src/main/java/com/arcadedb/query/sql/executor/ExpandStep.java
@@ -20,9 +20,11 @@ package com.arcadedb.query.sql.executor;
 
 import com.arcadedb.database.Document;
 import com.arcadedb.database.Identifiable;
+import com.arcadedb.database.MutableDocument;
 import com.arcadedb.database.Record;
 import com.arcadedb.exception.CommandExecutionException;
 import com.arcadedb.exception.TimeoutException;
+import com.arcadedb.security.AuthorizationUtils;
 
 import java.util.*;
 
@@ -70,12 +72,12 @@ public class ExpandStep extends AbstractExecutionStep {
           throw new NoSuchElementException();
 
         final Result result = nextElement;
+        
         localCount++;
         nextElement = null;
         fetchNext(context, nRecords);
         return result;
       }
-
     };
   }
 
@@ -92,6 +94,12 @@ public class ExpandStep extends AbstractExecutionStep {
             if (record == null) {
               continue;
             }
+         
+            // Prevent loading edges the user doesn't have access to.
+            if (!AuthorizationUtils.checkPermissionsOnDocument(record.asDocument(true), context.getDatabase())) {
+              continue;
+            }
+
             nextElement = new ResultInternal();
             ((ResultInternal) nextElement).setElement((Document) record);
           } else if (nextElementObj instanceof Map) {
