+++ engine/src/main/java/com/arcadedb/security/AuthorizationUtils.java
@@ -3,6 +3,10 @@ package com.arcadedb.security;
 import java.util.Map;
 import java.util.Set;
 
+import com.arcadedb.database.DatabaseInternal;
+import com.arcadedb.database.Document;
+import com.arcadedb.database.MutableDocument;
+
 public class AuthorizationUtils {
 
   /**
@@ -189,5 +193,68 @@ public class AuthorizationUtils {
         return classification.substring(0, classification.indexOf("//"));
     }
     return classification;
+  }
+
+
+  public static boolean checkPermissionsOnDocument(final Document document, final DatabaseInternal database) {
+
+    var currentUser = database.getContext().getCurrentUser();
+
+    // TODO short term - check classification, attribution on document
+
+    // TODO long term - replace with filtering by low classification of related/linked document.
+    // Expensive to do at read time. Include linkages and classification at write time?
+    // Needs performance testing and COA analysis.
+
+    if (currentUser.isServiceAccount() || currentUser.isDataSteward(document.getTypeName())) {
+      return true;
+    }
+
+    if ((!document.has(MutableDocument.CLASSIFICATION_MARKED) || !document.getBoolean(MutableDocument.CLASSIFICATION_MARKED))) {
+      return false;
+    }
+
+    // TODO detect and provide org for clearance
+    var clearance = currentUser.getClearanceForCountryOrTetragraphCode("USA");
+    var nationality = currentUser.getNationality();
+    var tetragraphs = currentUser.getTetragraphs();
+
+    if (document.has(MutableDocument.SOURCES)) {
+      // sources will be a map, in the form of source number : (classification//ACCM) source id
+      // check if user has appropriate clearance for any of the sources for the document
+      var isSourceAuthorized = document.toJSON().getJSONObject(MutableDocument.SOURCES).toMap().entrySet().stream().anyMatch(s -> {
+        
+        var source = s.getValue().toString();
+        if (source == null || source.isEmpty()) {
+          return false;
+        }
+
+        // if the source is not in the form of (classification/[/ACCM[]) [source id], then it is not a valid source
+        if (!source.contains("(") || !source.contains(")")) {
+          return false;
+        }
+
+        var sourceClassification = source.substring(source.indexOf("(") + 1, source.indexOf(")"));
+        return AuthorizationUtils.isUserAuthorizedForResourceMarking(clearance, nationality, tetragraphs, sourceClassification);
+      });
+
+      if (isSourceAuthorized) {
+        return true;
+      }
+    }
+
+    if (document.has(MutableDocument.CLASSIFICATION_PROPERTY) 
+          && document.toJSON().getJSONObject(MutableDocument.CLASSIFICATION_PROPERTY).has(MutableDocument.CLASSIFICATION_GENERAL_PROPERTY)) {
+      var docClassification = 
+          document.toJSON().getJSONObject(MutableDocument.CLASSIFICATION_PROPERTY).getString(MutableDocument.CLASSIFICATION_GENERAL_PROPERTY);
+      if (docClassification != null && !docClassification.isEmpty()) {
+        var isAuthorized = isUserAuthorizedForResourceMarking(clearance, nationality, tetragraphs, docClassification);
+        return isAuthorized;
+      } else {
+        return false;
+      }
+    }
+
+    return false;
   }
 }
